+++ src/Components/shared/Button/Button.js
@@ -1,14 +1,24 @@
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import './Button.scss';
 
-function Button({ label, className, icon, danger, transparent, cta, ...rest }) {
+function Button({
+	label,
+	className,
+	big,
+	icon,
+	danger,
+	transparent,
+	cta,
+	disabled,
+	...rest
+}) {
 	return (
 		<button
-			className={`button ${icon && 'button-icon'} ${
-				cta && 'button-cta'
-			} ${className} ${danger && 'button-danger'} ${
-				transparent && 'button-transparent'
-			}`}
+			className={`button ${disabled && 'button-disabled'} ${
+				big && 'button-big'
+			} ${icon && 'button-icon'} ${cta && 'button-cta'} ${className} ${
+				danger && 'button-danger'
+			} ${transparent && 'button-transparent'}`}
 			{...rest}>
 			{icon && <FontAwesomeIcon icon={icon} />}
 			{label}